Add a `template` property in AttributeItem
`template` is a readonly property that returns a string template of the
rst output. The return template is one of the four possible that are
defined as class attributes. The selection depends on the value (or lack
of value) of the DefinitionItem instance attrributes.

This greatly simplifies the to_rst method and I think it
improves readability. I might consider at a later time to expand this
behaviour to most of the DefinitionItem subclasses.